# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "族谱管理系统的RESTful API文档",
        "title": "族谱管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/familyRelation/create": {
            "post": {
                "description": "创建夫妻/伴侣关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "创建家庭关系",
                "parameters": [
                    {
                        "description": "家庭关系信息",
                        "name": "relation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/family_relation.FamilyRelation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/family_relation.FamilyRelation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/familyRelation/delete": {
            "delete": {
                "description": "删除夫妻/伴侣关系",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "删除家庭关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "家庭关系ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/familyRelation/getByFamilyId": {
            "get": {
                "description": "根据家庭ID获取其所有夫妻/伴侣关系",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "根据家庭ID获取家庭关系",
                "parameters": [
                    {
                        "type": "string",
                        "description": "家庭ID",
                        "name": "familyId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/family_relation.FamilyRelation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/familyRelation/getById": {
            "get": {
                "description": "根据ID获取夫妻/伴侣关系信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "根据ID获取家庭关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "家庭关系ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/family_relation.FamilyRelation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/familyRelation/getByPersonId": {
            "get": {
                "description": "根据人员ID获取其所有夫妻/伴侣关系",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "根据人员ID获取家庭关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人员ID",
                        "name": "personId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/family_relation.FamilyRelation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/familyRelation/getBySpouseId": {
            "get": {
                "description": "根据配偶ID获取其所有夫妻/伴侣关系",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "根据配偶ID获取家庭关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配偶ID",
                        "name": "spouseId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/family_relation.FamilyRelation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/familyRelation/update": {
            "put": {
                "description": "更新夫妻/伴侣关系信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "家庭关系管理"
                ],
                "summary": "更新家庭关系",
                "parameters": [
                    {
                        "description": "家庭关系信息",
                        "name": "relation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/family_relation.FamilyRelation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/family_relation.FamilyRelation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "family_relation.FamilyRelation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "divorce_date": {
                    "type": "string"
                },
                "family_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "marriage_date": {
                    "type": "string"
                },
                "marriage_place": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "person_id": {
                    "type": "integer"
                },
                "relation_type": {
                    "type": "string"
                },
                "spouse_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}